chore : 리뷰 카테고리 리스트 렌더링
