--- db.json
@@ -1,138 +1,0 @@
-{
-    "recentlyUsedBook": { 
-        "success": true,
-        "data": {
-          "pages": [
-            {
-              "id": 1,
-              "title": "청춘은 아름다워",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 2,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 3,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 4,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 5,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 6,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 7,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            },
-            {
-              "id": 8,
-              "title": "허상의 어릿광대",
-              "category": "소설",
-              "price": "10000",
-              "image": "https://img.ridicdn.net/cover/4653000052/xxlarge?dpi=xxhdpi#1"
-            }
-          ]
-        },
-        "error": null
-    },
-
-    "bestSeller": { 
-        "success": true,   
-        "data": {
-          "bookList" : [
-            {
-            "title": "월간순정 노자키 군 13 한정판 (82p 소책자 포함)",
-            "itemId": 285408606,
-            "cover": "https://image.aladin.co.kr/product/28540/86/coversum/k162835233_2.jpg",
-            "categoryName": "만화",
-            "bestRank": 1
-            },
-            {
-            "title": "The Earthian Tales 어션 테일즈 No.1 - alone",
-            "itemId": 284863481,
-            "cover": "https://image.aladin.co.kr/product/28486/34/coversum/k912835128_1.jpg",
-            "categoryName": "소설/시/희곡",
-            "bestRank": 2
-            },
-            {
-            "title": "설민석의 한국사 대모험 19 - 고려 편 : 왕건, 새로운 세상을 열다!",
-            "itemId": 285408840,
-            "cover": "https://image.aladin.co.kr/product/28540/88/coversum/k222835233_1.jpg",
-            "categoryName": "어린이",
-            "bestRank": 3
-            },
-            {
-            "title": "므레모사",
-            "itemId": 285549774,
-            "cover": "https://image.aladin.co.kr/product/28554/97/coversum/k162835148_1.jpg",
-            "categoryName": "소설/시/희곡",
-            "bestRank": 4
-            },
-            {
-            "title": "흔한남매 9",
-            "itemId": 283951526,
-            "cover": "https://image.aladin.co.kr/product/28395/15/coversum/k962835911_1.jpg",
-            "categoryName": "어린이",
-            "bestRank": 5
-            },
-            {
-            "title": "불편한 편의점",
-            "itemId": 269428498,
-            "cover": "https://image.aladin.co.kr/product/26942/84/coversum/k582730818_1.jpg",
-            "categoryName": "소설/시/희곡",
-            "bestRank": 6
-            },
-            {
-            "title": "한 게으른 시인의 이야기",
-            "itemId": 284079972,
-            "cover": "https://image.aladin.co.kr/product/28407/99/coversum/k162835916_2.jpg",
-            "categoryName": "에세이",
-            "bestRank": 7
-            },
-            {
-            "title": "마법천자문 52 - 불러 모아라! 모을 모(募)",
-            "itemId": 285273593,
-            "cover": "https://image.aladin.co.kr/product/28527/35/coversum/8950997711_1.jpg",
-            "categoryName": "어린이",
-            "bestRank": 8
-            },
-            {
-            "title": "웰씽킹 WEALTHINKING - 부를 창조하는 생각의 뿌리",
-            "itemId": 282733777,
-            "cover": "https://image.aladin.co.kr/product/28273/37/coversum/k852835990_3.jpg",
-            "categoryName": "자기계발",
-            "bestRank": 9
-            }
-          ]
-        },
-        "error": null
-  }
-}
