+++ src/components/Main/types.ts
@@ -1,0 +1,18 @@
+export interface BestSellerProps {
+  title: string;
+  categoryName: string;
+  cover: string;
+  bestRank: number;
+  link: string;
+  index: number;
+  isbn13: number;
+  isbn: number;
+}
+
+export interface BestSellerSliderProps {
+  id: number;
+  title: string;
+  category: string;
+  price: string;
+  image: string;
+}
