--- src/db/data.json
@@ -1,70 +1,0 @@
-{
-    "book": [
-          {
-            "id": 1,
-            "bookThumnail":"https://img.ridicdn.net/cover/1220000260/xxlarge#1?dpi=xxhdpi",
-            "bookName":"개정판 | 쓸 만한 인간",
-            "bookCategory":"에세이",
-            "authorName":"박정민",
-            "bookIntroTitle": "책소개",
-            "bookIntroText": "배우 박정민의 《쓸 만한 인간》이 3년여 만에 개정증보판으로 돌아왔다. 영화 〈파수꾼〉의 홍보용 블로그에 촬영 비하인드 스토리를 연재하면서 ‘글 좀 쓰는 배우’로 이름을 알렸던 그는 2013년부터 매거진 〈topclass〉에 칼럼을 실으며 독자층을 넓혀 갔다. ‘말로 기쁘게 한다’는 뜻의 언희(言喜)라는 필명처럼 재치 있는 필력과 유머러스한 이야기는 많은 사람들을 사로잡았고, 그 글을 모아 출간된 《쓸 만한 인간》은 ‘위로가 필요할 때마다 찾게 되는 책’ ‘나에게도, 누군가에게도 너무나 좋은 선물이 될 책’ ‘유쾌하지만 결코 가볍지 않다’ 등 공감과 위로를 자아내며 많은 이들에게 사랑받았다. 여기에 새 글과 일러스트를 더해 3년 만에 개정증보판으로 출간됐다. 이번 개정증보판에서 가장 눈에 띄는 변화는 박정민이 직접 쓰고 그린 일러스트와 손글씨가 추가됐다는 점이다. ‘글을 말로 옮기는 일을 하다가 말을 글로 옮기고 싶어졌다’고 말했던 그가 이번에는 일러스트까지 더해 좀 더 솔직한 자신의 이야기를 담았다. 또한 작가 특유의 유머러스한 문체를 유지하면서도 시간이 흐름을 반영하기 위해 문장을 꼼꼼하게 다시 들여다보며 수정하는 작업을 거쳤다. 기존 도서에는 실리지 않았던 〈topclass〉 연재분은 물론, 최근의 기록까지 총 9편의 글도 새롭게 수록됐다.",
-            "authorIntroTitle": "저자 소개",
-            "authorIntroText":
-              "1987년 충주에서 태어나 가히 모범적이라 할 수 있는 학창 시절을 보냈다. 2005년 고려대학교 인문학부에 입학했으나 연기를 ‘제대로’ 하고 싶다는 열망으로 자퇴를 했다. 이후 한국예술종합학교에 입학해 몇 편의 단편영화를 거쳐, 2011년 독립영화 〈파수꾼〉으로 데뷔했다. 〈전설의 주먹〉 〈동주〉 〈그것만이 내 세상〉 〈변산〉 〈사바하〉 〈타짜: 원 아이드 잭〉 등의 영화에 출연하였고, 〈너희들은 포위됐다〉 〈안투라지〉 등의 드라마와 〈키사라기 미키짱〉 〈 G코드의 탈출〉 등의 연극 무대에도 올랐다. ‘말로 기쁘게 한다.’는 뜻의 언희(言喜)라는 필명으로 2013년부터 매거진 〈topclass〉에 칼럼을 연재하기도 했다. "
-          }
-    ],
-    "img": [
-          {
-            "id": 1,
-            "img": "https://img.ridicdn.net/cover/222002478/xlarge"
-          },
-          {
-            "id": 2,
-            "img": "https://img.ridicdn.net/cover/606002187/xlarge"
-          },
-          {
-            "id": 3,
-            "img": "https://img.ridicdn.net/cover/734001913/xlarge"
-          },
-          {
-            "id": 4,
-            "img": "https://img.ridicdn.net/cover/2189000171/xxlarge"
-          },
-          {
-            "id": 5,
-            "img": "https://img.ridicdn.net/cover/222001951/xlarge"
-          }
-    ],
-    "comments": [
-        {
-          "bookId":1314000079,
-          "user": 1,
-          "content": "읽기 부담없고 술술 잘읽혀요 유머가 유쾌해요. 서른 앞두고 고민이 많았는데 배우 박정민도 고민이 많았었구나 다들 이렇게 사는구나 느꼈네요.",
-          "replyDate": "2021-12-22"
-        },
-        {
-          "bookId":1314000079,
-          "user": 2,
-          "content": "글에서도 음성지원이 되는듯! 글을 읽지만 소리로 듣는 느낌ㅎㅎ",
-          "replyDate": "2021-12-19"
-        },
-        {
-          "bookId":1314000079,
-          "user": 3,
-          "content": "읽는 동안 박정민이라는 배우 캐릭터가 눈앞에 아른거리며 글에서 느껴지는 유머가 책 읽기 싫어하는 나를 손 못놓게 했음. 재밌습니다.",
-          "replyDate": "2021-11-30"
-        },
-        {
-          "bookId":1314000079,
-          "user": 4,
-          "content": "서점에서 잠깐 읽어보고 마음에 와닿는 구절들이 많아 이북으로 구매 했어요! 책을 읽는 동안 박정민이라는 사람의 유쾌함과 배우로써 느끼는 여러가지 고뇌들을 조금이나마 엿볼 수 있는 책이였어요. 서점에서 잠깐 읽어보고 마음에 와닿는 구절들이 많아 이북으로 구매 했어요! 책을 읽는 동안 박정민이라는 사람의 유쾌함과 배우로써 느끼는 여러가지 고뇌들을 조금이나마 엿볼 수 있는 책이였어요.",
-          "replyDate": "2021-10-27"
-        },
-        {
-          "bookId":1314000079,
-          "user": 5,
-          "content": "보고싶어서 기다리다가 이북 나오자마자 구매해서 읽었습니다. 시간가는줄모르고 아주 낄낄거리면서 읽을 수있는 신박한 에세이 입니다 강추요",
-          "replyDate": "2021-09-11"
-        }
-      ]
-    }
