+++ src/modules/Slices/book/bookSlice.ts
@@ -2,93 +2,44 @@ import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import { AxiosError } from "axios";
 import { RootState } from "modules/store";
 import http from "src/api/http";
-
-// initialState로 사용할 객체의 data type
-interface BookItemProps {
-  author: string;
-  isbn: string;
-  link: string;
-  description: string;
-  title: string;
-  pubDate: string;
-  categoryName: string;
-  fixedPrice: boolean;
-  mallType: string;
-  customerReviewRank: number;
-  cover: string;
-  itemId: number;
-  subInfo: any;
-  isbn13: string;
-  stockStatus: string;
-  publisher: string;
-  priceSales: number;
-  salesPoint: number;
-  adult: boolean;
-  categoryId: number;
-  priceStandard: number;
-  mileage: number;
-  bestRank: number;
-  // itemId: number;
-  // title: string;
-  // categoryName: string;
-  // cover: string;
-  // bestRank: number;
-}
-
-// 리듀가 사용할 데이터 타입
-interface BookListReduceProps {
-  item: BookItemProps[];
-  status: "loading" | "idle";
-  error: null | {
-    code: number;
-    message: string;
-  };
-}
-
-// 썽크함수 성공시 반환 타입
-interface getBookAsyncSucess {
-  success: boolean;
-  data: {
-    totalResults: number;
-    startIndex: number;
-    item: BookItemProps[];
-  };
-  error: null;
-}
-
-// 썽크함수가 실패시 반환 타입
-interface getBookAsyncFail {
-  success: boolean;
-  data: null;
-  error: {
-    code: number;
-    message: string;
-  };
-}
-
-// 썽크함수가 사용하는 api 타입
-interface ThunkApi {
-  rejectValue: getBookAsyncFail;
-}
+import { BookListReduceProps, getBookAsyncFail, getBookAsyncSuccess } from "./types";
 
 const name = "bookReduce";
 
-// Middleware
-export const getbookAPI = createAsyncThunk<getBookAsyncSucess, undefined, ThunkApi>(
-  `${name}/bookAsync`,
+const initialState: BookListReduceProps = {
+  item: [],
+  status: "loading",
+  error: null,
+};
+
+// 베스트셀러
+export const getBookAPI = createAsyncThunk<getBookAsyncSuccess>(`${name}/bookAsync`, async (_, { rejectWithValue }) => {
+  try {
+    const response = await http.get("/book/bestseller?page=1&size=9");
+    return response.data;
+  } catch (err) {
+    const error = err as AxiosError<getBookAsyncFail>;
+    if (!error.response) throw err;
+    return rejectWithValue(error.response.data);
+  }
+});
+
+// 카테고리
+export const getCategoryBook = createAsyncThunk<getBookAsyncSuccess>(
+  `getCategoryBook/bookAsync`,
   async (_, { rejectWithValue }) => {
     try {
-      const response = await http.get("/book/bestseller?page=1&size=9");
+      const response = await http.get(`/book/byCategory?categoryId=74&page=0&size=20`);
       return response.data;
     } catch (err) {
       const error = err as AxiosError<getBookAsyncFail>;
-      if (!error.response) throw err; // 런타임 에러
+      if (!error.response) throw err;
       return rejectWithValue(error.response.data);
     }
   },
 );
 
-// export const getbookAPItest = createAsyncThunk<getBookAsyncSucess, BookItemProps, ThunkApi>(
+// export const getBookAPItest = createAsyncThunk<getBookAsyncSucess, BookItemProps, ThunkApi>(
 //   `${name}/bookAsync`,
 //   async ({ itemId, title, categoryName, cover, bestRank }, { rejectWithValue }) => {
 //     try {
@@ -106,41 +57,54 @@ export const getbookAPI = createAsyncThunk<getBookAsyncSucess, undefined, ThunkA
 // Slice
 export const getBookSlice = createSlice({
   name: "bookList",
-  initialState: {
-    item: [
-      // {
-      //   itemId: 12345,
-      //   title: "책이름",
-      //   categoryName: "소설",
-      //   cover: "https://image.aladin.co.kr/product/28540/88/coversum/k222835233_1.jpg",
-      //   bestRank: 1,
-      // },
-    ],
-    status: "idle",
-    error: null,
-  } as BookListReduceProps,
+  initialState,
   reducers: {},
   extraReducers: builder => {
-    builder.addCase(getbookAPI.pending, state => {
-      state.error = null;
-      state.status = "loading";
-    });
-    builder.addCase(getbookAPI.fulfilled, (state, { payload }) => {
-      console.log(payload);
-      state.item = payload.data.item;
-    });
-    builder.addCase(getbookAPI.rejected, (state, { payload }) => {
-      console.log("error");
-      // 에러핸들링
-      if (!payload) {
-        state.error = {
-          code: 500,
-          message: "서버에서 데이터 못가져옴",
-        };
-      }
-    });
+    builder
+      // 베스트셀러
+      .addCase(getBookAPI.pending, state => {
+        state.error = null;
+        state.status = "loading";
+      })
+      .addCase(getBookAPI.fulfilled, (state, { payload }) => {
+        state.item = payload.data.item;
+        state.status = "idle";
+      })
+      .addCase(getBookAPI.rejected, (state, { payload }) => {
+        console.log("error");
+        state.status = "loading";
+        // 에러핸들링
+        if (!payload) {
+          state.error = {
+            code: 500,
+            message: "서버에서 데이터 못가져옴",
+          };
+        }
+      })
+      // 카테고리 리스트
+      .addCase(getCategoryBook.pending, state => {
+        state.error = null;
+        state.status = "loading";
+      })
+      .addCase(getCategoryBook.fulfilled, (state, { payload }) => {
+        state.item = payload.data.item;
+        state.status = "idle";
+      })
+      .addCase(getCategoryBook.rejected, (state, { payload }) => {
+        console.log("error");
+        state.status = "loading";
+        // 에러핸들링
+        if (!payload) {
+          state.error = {
+            code: 500,
+            message: "서버에서 데이터 못가져옴",
+          };
+        }
+      });
   },
 });
 
 export const getBookSelector = (state: RootState) => state.bookListReduce;
+export const getBookItemList = (state: RootState) => state.bookListReduce.item;
+
 export default getBookSlice;
